pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ds-oracle"
version = "1.0.0"
description = "DS Oracle 판정 엔진 서버 (Rule-based 1차 검증)"
requires-python = ">=3.11"
authors = [{ name = "DS Capstone Team" }]
dependencies = [
    "paho-mqtt>=2.1.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
ds-oracle = "main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
